Add overrides for `stargate.debug.enabled` for test mode; move config class
