+++ src/main/java/io/stargate/sgv2/jsonapi/config/DebugModeConfig.java
@@ -1,4 +1,4 @@
-package io.stargate.sgv2.jsonapi.exception;
+package io.stargate.sgv2.jsonapi.config;
 
 import io.smallrye.config.ConfigMapping;
 import io.smallrye.config.WithDefault;
@@ -7,5 +7,5 @@ import io.smallrye.config.WithDefault;
 public interface DebugModeConfig {
 
   @WithDefault("false")
-  boolean enable();
+  boolean enabled();
 }
